pyproject.toml
[project]
name = "gcash-reviews-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dbt-bigquery>=1.11.1",
    "dbt-duckdb>=1.10.1",
    "duckdb>=1.5.2",
    "google-cloud-storage>=2.18.0",
    "google-play-scraper>=1.2.7",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
    "streamlit>=1.56.0",
    "wordcloud>=1.9.6",
]


[tool.uv]
package = true  # ← add this

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ingestion", "processing"]
